Harden native OAuth callback handling and isolate URL builder
